Add data processing scripts for loading, cleaning, merging, and transforming data
