+++ notebooks/scripts/save_data.py
@@ -1,0 +1,4 @@
+def save_to_csv(df, file_path):
+    """Save the DataFrame to a CSV file."""
+    df.to_csv(file_path, index=False)
+
